+++ src/main/java/balancetalk/member/application/MemberService.java
@@ -6,13 +6,13 @@ import balancetalk.global.exception.BalanceTalkException;
 import balancetalk.global.exception.ErrorCode;
 import balancetalk.global.jwt.JwtTokenProvider;
 import balancetalk.global.redis.application.RedisService;
-import balancetalk.member.dto.JoinRequest;
-import balancetalk.member.dto.LoginRequest;
-import balancetalk.member.dto.MemberResponse;
 import balancetalk.file.domain.File;
 import balancetalk.file.domain.FileRepository;
 import balancetalk.member.domain.Member;
 import balancetalk.member.domain.MemberRepository;
+import balancetalk.member.dto.MemberDto.JoinRequest;
+import balancetalk.member.dto.MemberDto.LoginRequest;
+import balancetalk.member.dto.MemberDto.MemberResponse;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
